step1: save() 인자 검증 이외 비즈니스 로직 Service로 이전 모든 메소드 사용자 이름 변수 전달 파일 다운로드 기능 추가

+++ Back-EndProject/src/main/java/org/example/mvc/controller/UploadFileController.java
@@ -3,12 +3,13 @@ package org.example.mvc.controller;
 import org.example.configuration.GlobalConfig;
 import org.example.configuration.exception.BaseException;
 import org.example.configuration.http.BaseResponseCode;
-import org.example.mvc.domain.dto.MetadataInsertDTO;
+import org.example.mvc.domain.dto.MetaDataDTO;
 import org.example.mvc.domain.dto.UploadFileDTO;
 import org.example.mvc.service.UploadFileService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -16,6 +17,8 @@ import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
 import java.io.IOException;
+import java.io.UnsupportedEncodingException;
+import java.net.URLEncoder;
 import java.util.List;
 import java.util.UUID;
 
@@ -30,63 +33,55 @@ public class UploadFileController {
     private UploadFileService uploadFileService;
 
     @RequestMapping("/save")
-    public String save(@RequestParam("uploadFile") MultipartFile multipartFile, @RequestParam("userName") String userName) throws IOException{
+    public String save(@RequestParam("uploadFile") MultipartFile multipartFile, @RequestParam("userName") String userName, Model model) throws IOException{
+
         if(multipartFile == null || multipartFile.isEmpty()){
-            throw new BaseException(BaseResponseCode.DATA_IS_NULL);
+            throw new BaseException(BaseResponseCode.DATA_IS_NULL, new String[]{"업로드 파일"});
         }
 
         if(userName.isEmpty()){
             throw new BaseException(BaseResponseCode.DATA_IS_NULL, new String[]{"사용자 이름"});
         }
 
-        String uploadFilePath = globalConfig.getUploadFilePath();
-        String prefix = multipartFile.getOriginalFilename().substring(multipartFile.getOriginalFilename().lastIndexOf(".") + 1, multipartFile.getOriginalFilename().length());
+        logger.debug("save : {}",userName);
+        model.addAttribute("userName",userName);
 
-        boolean thumbnailCheck = false;
-        if(prefix.equals("jpg") || prefix.equals("png") || prefix.equals("jpeg") || prefix.equals("gif")) thumbnailCheck = true;
-
-        String filename = UUID.randomUUID().toString() + "." +  prefix;
-
-        File folder = new File(uploadFilePath);
-
-        if(!folder.isDirectory()){
-            folder.mkdirs();
+        String encodedUserName;
+        try {
+            encodedUserName = URLEncoder.encode(userName, "UTF-8");
+        } catch (UnsupportedEncodingException e) {
+            throw new RuntimeException("Error encoding userName", e);
         }
 
-        String pathname = uploadFilePath + filename;
-
-        try{
-            MetadataInsertDTO parameter = new MetadataInsertDTO();
-            parameter.setUserName(userName);
-            parameter.setFileType(false);
-            parameter.setSize((int)multipartFile.getSize());
-            parameter.setFileName(multipartFile.getOriginalFilename().substring(0,multipartFile.getOriginalFilename().lastIndexOf(".")));
-            parameter.setFilePath(uploadFilePath + multipartFile.getOriginalFilename());
-            parameter.setOriginalPath(pathname);
-            parameter.setThumbnailCheck(thumbnailCheck);
-            uploadFileService.save(parameter);
-
-            File dest = new File(pathname);
-            multipartFile.transferTo(dest);
-        }catch (IllegalStateException | IOException e){
-            logger.error("e",e);
-        }
-        return "redirect:/file/getList";
+        uploadFileService.fileSave(multipartFile,userName);
+        return "redirect:/file/getList?userName=" + encodedUserName;
     }
 
     @GetMapping("/getList")
-    public String getUploadFileList(Model model){
-        List<UploadFileDTO> list = uploadFileService.getUploadFileList();
+    public String getList(@RequestParam(name = "userName",required = false)String userName,Model model){
+        List<MetaDataDTO> list = uploadFileService.getList();
+        model.addAttribute("userName",userName);
         model.addAttribute("fileList",list);
         return "uploadFile/file_list.html";
     }
 
     @PostMapping("/delete/{fileSeq}")
     public String delete(@PathVariable int fileSeq){
-
+        if(fileSeq < 0){
+            throw new BaseException(BaseResponseCode.VALIDATE_REQUIRED,new String[]{"파일 번호"});
+        }
 
         uploadFileService.delete(fileSeq);
         return "redirect:/file/getList";
+    }
+
+    @GetMapping("/download/{fileSeq}")
+    public ResponseEntity<byte[]> downloadFile(@PathVariable int fileSeq) throws IOException {
+        if(fileSeq < 0){
+            throw new BaseException(BaseResponseCode.VALIDATE_REQUIRED,new String[]{"파일 번호"});
+        }
+
+        return uploadFileService.downloadFile(fileSeq);
     }
 
 }
